# desktop_app/openapi/archestra/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Archestra API",
    "version": "0.0.1"
  },
  "components": {
    "schemas": {
      "OAuthTokensInput": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "number"
          },
          "token_type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        },
        "required": ["access_token"],
        "additionalProperties": {}
      },
      "OAuthClientInformationInput": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          }
        },
        "required": ["client_id"],
        "additionalProperties": {}
      },
      "AuthorizationServerMetadataInput": {
        "type": "object",
        "properties": {
          "issuer": {
            "type": "string"
          },
          "authorization_endpoint": {
            "type": "string"
          },
          "token_endpoint": {
            "type": "string"
          },
          "revocation_endpoint": {
            "type": "string"
          },
          "scopes_supported": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": {}
      },
      "OAuthProtectedResourceMetadataInput": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string"
          },
          "scopes_supported": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": {}
      },
...
      "McpServerConfigInput": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": ["command"],
        "additionalProperties": false
      },
      "McpServerUserConfigValuesInput": {
        "type": "object",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "number"
            },
            {
              "type": "boolean"
            },
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          ]
        }
      },
...
      "McpServerInstallInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "pattern": "^[A-Za-z0-9-\\s]{1,63}$"
          },
          "serverConfig": {
            "$ref": "#/components/schemas/McpServerConfigInput"
          },
          "userConfigValues": {
            "$ref": "#/components/schemas/McpServerUserConfigValuesInput"
          },
          "oauthProvider": {
            "type": "string"
          },
          "oauthTokens": {
            "$ref": "#/components/schemas/OAuthTokensInput"
          },
          "oauthClientInfo": {
            "$ref": "#/components/schemas/OAuthClientInformationInput"
          },
          "oauthServerMetadata": {
            "$ref": "#/components/schemas/AuthorizationServerMetadataInput"
          },
          "oauthResourceMetadata": {
            "$ref": "#/components/schemas/OAuthProtectedResourceMetadataInput"
          }
        },
        "required": ["displayName", "serverConfig"]
      },
...
      "OAuthTokens": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "number"
          },
          "token_type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        },
        "required": ["access_token"],
        "additionalProperties": {}
      },
      "OAuthClientInformation": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": "string"
          }
        },
        "required": ["client_id"],
        "additionalProperties": {}
      },
      "AuthorizationServerMetadata": {
        "type": "object",
        "properties": {
          "issuer": {
            "type": "string"
          },
          "authorization_endpoint": {
            "type": "string"
          },
          "token_endpoint": {
            "type": "string"
          },
          "revocation_endpoint": {
            "type": "string"
          },
          "scopes_supported": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": {}
      },
      "OAuthProtectedResourceMetadata": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string"
          },
          "scopes_supported": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": {}
      },
...
      "McpServerConfig": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "args": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": ["command"],
        "additionalProperties": false
      },
      "McpServerUserConfigValues": {
        "type": "object",
        "additionalProperties": {
          "anyOf": [
            {
              "type": "string"
            },
            {
              "type": "number"
            },
            {
              "type": "boolean"
            },
            {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          ]
        }
      },
      "McpServer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "serverConfig": {
            "$ref": "#/components/schemas/McpServerConfig"
          },
          "userConfigValues": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpServerUserConfigValues"
              },
              {
                "nullable": true,
                "enum": [null]
              }
            ]
          },
          "oauthTokens": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OAuthTokens"
              },
              {
                "nullable": true,
                "enum": [null]
              }
            ]
          },
          "oauthClientInfo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OAuthClientInformation"
              },
              {
                "nullable": true,
                "enum": [null]
              }
            ]
          },
          "oauthServerMetadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorizationServerMetadata"
              },
              {
                "nullable": true,
                "enum": [null]
              }
            ]
          },
          "oauthResourceMetadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OAuthProtectedResourceMetadata"
              },
              {
                "nullable": true,
                "enum": [null]
              }
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "serverConfig",
          "userConfigValues",
          "oauthTokens",
          "oauthClientInfo",
          "oauthServerMetadata",
          "oauthResourceMetadata",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
      "McpServerContainerLogs": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "string"
          },
          "containerName": {
            "type": "string"
          }
        },
        "required": ["logs", "containerName"],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/mcp_server/install": {
      "post": {
        "operationId": "installMcpServer",
        "tags": ["MCP Server"],
        "description": "Install an MCP server. Either from the catalog, or a customer server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerInstallInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServer"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp_server/{id}": {
      "delete": {
        "operationId": "uninstallMcpServer",
        "tags": ["MCP Server"],
        "description": "Uninstall MCP server",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": ["success"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/mcp_proxy/{id}/logs": {
      "get": {
        "operationId": "getMcpServerLogs",
        "tags": ["MCP Server"],
        "description": "Get logs for a specific MCP server container",
        "parameters": [
          {
            "schema": {
              "default": 100,
              "type": "number"
            },
            "in": "query",
            "name": "lines",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerContainerLogs"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mcp_server/oauth_install": {
      "post": {
        "operationId": "installMcpServerWithOauth",
        "tags": ["MCP Server"],
        "description": "Install MCP server with OAuth authentication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string"
                  },
                  "installData": {
                    "$ref": "#/components/schemas/McpServerInstallInput"
                  }
                },
                "required": ["provider", "installData"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "server": {
                      "$ref": "#/components/schemas/McpServer"
                    }
                  },
                  "required": ["server"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}