# desktop_app/src/backend/database/schema/mcpServer.ts
import { sql } from 'drizzle-orm';
import { sqliteTable, text } from 'drizzle-orm/sqlite-core';
import { z } from 'zod';

/**
 * MCP Server Status Enum
 */
export const McpServerStatusSchema = z.enum(['installing', 'oauth_pending', 'installed', 'failed']);
export type McpServerStatus = z.infer<typeof McpServerStatusSchema>;

import {
  type AuthorizationServerMetadata,
  AuthorizationServerMetadataSchema,
  type OAuthClientInformation,
  OAuthClientInformationSchema,
  type OAuthProtectedResourceMetadata,
  OAuthProtectedResourceMetadataSchema,
  type OAuthTokens,
  OAuthTokensSchema,
} from './oauth';

/**
 * Borrowed from @anthropic-ai/dxt
 *
 * https://github.com/anthropics/dxt/blob/v0.2.6/src/schemas.ts#L3-L7
 */
export const McpServerConfigSchema = z.strictObject({
  command: z.string(),
  args: z.array(z.string()).optional(),
  env: z.record(z.string(), z.string()).optional(),
});

export const McpServerUserConfigValuesSchema = z.record(
  z.string(),
  z.union([z.string(), z.number(), z.boolean(), z.array(z.string())])
);

export const mcpServersTable = sqliteTable('mcp_servers', {
  /**
   * Catalog "name" (unique identifier) or UUID for custom servers
   */
  id: text().primaryKey(),
  /**
   * Display name (from catalog or user-defined for custom)
   */
  name: text().notNull(),
  /**
   * https://orm.drizzle.team/docs/column-types/sqlite#blob
   */
  serverConfig: text({ mode: 'json' }).$type<z.infer<typeof McpServerConfigSchema>>().notNull(),
  /**
   * `userConfigValues` are user-provided/custom values for `DxtManifestMcpConfig`
   * (think API keys, etc)
   *
   * This is only used for mcp servers installed via the catalog, as it allows users to provide
   * dynamic configuration
   *
   * See https://github.com/anthropics/dxt/blob/main/MANIFEST.md#variable-substitution-in-user-configuration
   */
  userConfigValues: text({ mode: 'json' }).$type<z.infer<typeof McpServerUserConfigValuesSchema>>(),
  /**
   * OAuth tokens object - matches OAuthTokens interface from MCP SDK
   * Stores access_token, refresh_token, expires_in, token_type, scope, etc.
   */
  oauthTokens: text('oauth_tokens', { mode: 'json' }).$type<OAuthTokens>(),
  /**
   * OAuth client information - matches OAuthClientInformation interface from MCP SDK
   * Stores client_id, client_secret (for static registration or dynamic registration result)
   */
  oauthClientInfo: text('oauth_client_info', { mode: 'json' }).$type<OAuthClientInformation>(),
  /**
   * OAuth authorization server metadata - matches AuthorizationServerMetadata interface from MCP SDK
   * Stores OAuth discovery metadata per RFC 8414 (issuer, endpoints, scopes_supported, etc.)
   */
  oauthServerMetadata: text('oauth_server_metadata', { mode: 'json' }).$type<AuthorizationServerMetadata>(),
  /**
   * OAuth protected resource metadata - matches OAuthProtectedResourceMetadata interface from MCP SDK
   * Stores resource-specific OAuth metadata (resource identifier, required scopes, etc.)
   */
  oauthResourceMetadata: text('oauth_resource_metadata', { mode: 'json' }).$type<OAuthProtectedResourceMetadata>(),
  /**
   * Current status of the MCP server installation/OAuth flow
   */
  status: text().notNull().default('installing').$type<McpServerStatus>(),
  createdAt: text()
    .notNull()
    .default(sql`(current_timestamp)`),
});

/**
 * Pure Zod schema for OpenAPI generation
 * This matches the structure of the database table but uses pure Zod types
 */
export const McpServerSchema = z.object({
  id: z.string(),
  name: z.string(),
  serverConfig: McpServerConfigSchema,
  userConfigValues: McpServerUserConfigValuesSchema.nullable(),
  oauthTokens: OAuthTokensSchema.nullable(),
  oauthClientInfo: OAuthClientInformationSchema.nullable(),
  oauthServerMetadata: AuthorizationServerMetadataSchema.nullable(),
  oauthResourceMetadata: OAuthProtectedResourceMetadataSchema.nullable(),
  status: McpServerStatusSchema,
  createdAt: z.string(),
});

export type McpServer = z.infer<typeof McpServerSchema>;
export type McpServerConfig = z.infer<typeof McpServerConfigSchema>;
export type McpServerUserConfigValues = z.infer<typeof McpServerUserConfigValuesSchema>;